move EEROS and EEORS flink to production image
The current way to build EEROS sets LIBUCL_INCLUDE_DIR and LIBUCL_LIB_DIR based on the build/output path for cmake.
This results in build errors when trying to build EEROS applications when EEROS is included in the SDK,
since those paths point to yocto's build/tmp/work/... directory, which probably does not exist when using the SDK.
Therefore it is not currently possible to include EEROS in the image and usse the build scripts to build an EEROS application.
It should however still work when everything (including the application itself) is built with yocto,
since everything is built in build/tmp/work/... . Hence it is still incuded in the prod image.

Building EEROS for the dev image needs to be done using the EEROS build scripts as before.